Tidy up test utils
