+++ ui/core/src/test/utils/accounts.ts
@@ -1,0 +1,6 @@
+export type TestSifAccount = { mnemonic: string; address: string };
+export const akasha = {
+  mnemonic:
+    "race draft rival universe maid cheese steel logic crowd fork comic easy truth drift tomorrow eye buddy head time cash swing swift midnight borrow",
+  address: "sif1l7hypmqk2yc334vc6vmdwzp5sdefygj2ad93p5",
+};
